Add nowide library, begin using std::string more
nowide is a library to add utf8 support to standard functions on
Windows. On other platforms it simply behaves as an alias. This should
simplify and stabalise our unicode code.

We should start using STL types and structures more; we are using C++11
after all! So I've begun transitioning some of the newer methods over to
std::string.

ShowError() now also prints to stderr - this is especially useful if the
error message has characters that are missing from the ImGui font.