CMake: Don't set CMAKE_INSTALL_PREFIX internally,
use wrapper script instead.

Common use case is to pass DESTDIR as a destination directory,
and CMAKE_INSTALL_PREFIX as a prefix like for example:

cmake -DCMAKE_INSTALL_PREFIX=/usr
...
make install DESTDIR=/buildroot/

As it was, this resulted in an unwanted prefix, and there was no
way of controlling the prefix externally.