Added SDL2-devel-2.0.5-VC into SDC2/win32 directory
Added the include files and libs for the Windows version of SDC2, also
modified the CMakeLists.txt file to allow MSVC to compile with the
files.
Ideally it would be better if these files are pulled from the SDC server
but alas I am unable to do such things, maybe someone else can in the
future.