Revert "Reduce travis clone depth to 5"
This reverts commit 57e78441cd1c20d1644c34de188243b4aa45558b.

We use the number of commits and the current branch/tag to identify a specific build (with the `OBV_BUILD` macro), so we need the full Git history.
This means a shallow clone won't work, and that's why Travis' depth is set to 9999999.
I don't have a better solution for now, it works and the shallow clone is only 47% lighter than a full clone so that's not a huge difference (considering our repo is less than 15MiB).