Fix memory leak happening when opening file
The issue was caused by stringfile() allocating memory for lines pointers, and
this memory was never freed (even though the comment of the function explicitly
mentioned that the result is to be freed).

This commit replaces allocation of bare memory with using std::vector to store
pointers to all lines. This saves us a trouble of manually freeing memory.

Most of the board parsers became shorter, but some of them are still using
bare pointers internally because they need to pass line pointer around and
modify it (i.e. skip lines). It is probably possible to make that code more
modern, but could happen as a separate change.