The new "Including details" checkbox is added to search dialog.
It enables searching nets by pin names/numbers and
components by the "mfgcode" string.

For me it is useful for finding pins by name on large bga arrays
and answering questions
"Where did they put that P-mosfet in this revision of design?"
(boards I'm working with has a lot of mfgcode data - part number, Ohm values, packages).

Checkbox is disabled by default, since it is slower on huge boards and
UI is a bit less expected then other searches -
the results that are showing below search fields are still pin/net names,
so it isn't immediately obvious why they are included there.

Also since entire net is found in by-pin-name search -
it is actually not applicable for determining for exact pin location
for nets that have lots of associated pins (GND, power, etc).