CMake: Search for stb through pkg-config, with find_path() fallback
When not targeting Windows, pkg-config is asked for the package.

The problem is that the ".pc" file is distro-specific, as it's not upstreamed.
As a solution this commit implements a fallback using find_path().
Hints can be provided through the "STB_INCLUDE_DIRS" variable.

Finally, when everything else fails, we resort to the submodule.